% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_domestic_product}
\alias{growth_domestic_product}
\title{Growth Domestic Product}
\format{
A data frame with 101 rows and 28 columns:
\describe{
  \item{country}{the country in which growth domestic product was measured.}
  \item{1990}{growth domestic product in 1990 for each country.}
  ...
  \item{2016}{growth domestic product in 2016 for each country}
}
}
\source{
\url{https://www.imf.org/external/datamapper/NGDPD@WEO/OEMDC/ADVEC/WEOWORLD}
}
\usage{
growth_domestic_product
}
\description{
A dataset containing the growth domestic product for several countries from 1990 to 2016.
}
\keyword{datasets}
