\name{idld-package}
\alias{idld-package}
\alias{idld}
\docType{package}
\title{
\packageTitle{idld}
}
\description{
\packageDescription{idld}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{idld}
\packageIndices{idld}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{idld}

Maintainer: \packageMaintainer{idld}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
