% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idld_cluster.R
\name{idld_cluster}
\alias{idld_cluster}
\title{IDLD Clustering}
\arguments{
\item{Z}{data to apply depth. It should be an array of dimension (n,p,l) where l is the number of functional coordinates. 
Z[,,i] is a numeric matrix where each row represents a functional observation for i=1,...,l.}

\item{data_mf}{data on which depth is based. Same format than Z.}

\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{alpha_quantile}{proportions of data points to include in the deepest regions. It could be a numeric vector.}

\item{K}{number of clusters}

\item{type}{the data type to apply the idld, "multivariate", "functional" or "multi_functional".}

\item{verbose}{if TRUE prints the algorithm progress.}
}
\value{
returns a list with the following components:
\itemize{  
\item local_depth: A numeric vector object that contains the depth for each point.
\item region: A matrix containing, in each column, the data which is in the central region related to alpha_quantile selected. 
\item clusters a matrix containing, in each column, the data partition related to alpha_quantile selected.
}
}
\description{
It is partition-based clustering technique based on local depth and distance measurement applied to data
}
