% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idld_cluster_boot.R
\name{idld_cluster_boot}
\alias{idld_cluster_boot}
\title{IDLD Clustering with Bootstrap}
\usage{
idld_cluster_boot(Z, beta, m, K, B, type, verbose = FALSE)
}
\arguments{
\item{Z}{data to apply depth. It should be an array of dimension (n,p,l) where l is the number of functional coordinates. 
Z[,,i] is a numeric matrix where each row represents a functional observation for i=1,...,l.}

\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{K}{number of clusters}

\item{B}{number of bootstrap samples}

\item{type}{the data type to apply the idld, "multivariate", "functional" or "multi_functional".}

\item{verbose}{if TRUE prints the algorithm progress.}
}
\value{
returns a list with the following components: 
\itemize { 
\item local_depth a numeric vector object that contains the depth for each point.
\item region vector of booleans indicating which data points is in the central region. 
\item clusters numeric vector with the partition.
}
}
\description{
It is partition-based clustering technique based on local depth and distance measurement applied to data.
This functional selects the optimal alpha_quantile for the procedure.
}
\examples{
library(funHDDC)
library(abind)
data("triangle")
triangle_data = abind(triangle[,1:101], triangle[,102:202], along=3)
d = dim(triangle_data)
triang_cl = idld_cluster_boot(triangle_data, 0.2, 100, 3, 20, "multi_functional", TRUE)
par(mfrow=c(1,2))
plot(triangle_data[1,,1], type="n", ylim=c(0,8))
for (i in 1:d[1]) lines(triangle_data[i,,1], col=triang_cl$clusters[i])
plot(triangle_data[1,,2], type="n", ylim=c(0,8))
for (i in 1:d[1]) lines(triangle_data[i,,2], col=triang_cl$clusters[i])

}
