% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idld_functional.R
\name{idld_f}
\alias{idld_f}
\title{Functional Integrated Dual Local Depth}
\arguments{
\item{Z}{data to apply depth. It should be a numeric matrix where each row represents an observation.}

\item{data_f}{data on which depth is based. It should be a numeric matrix where each row represents an observation.}

\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{verbose}{if TRUE prints the algorithm progress.}
}
\value{
A numeric vector object that contains the depth for each point.
}
\description{
Calculates the integrated dual local depth for functional data and multivariate functional data.
}
\examples{
data(inflation_rate)
inflation = as.matrix(inflation_rate[,-1])
local_depth = idld_f(inflation, inflation, 0.3, 500, TRUE)
}
