% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idld.R
\name{idld_function}
\alias{idld_function}
\alias{idld}
\title{Integrated Dual Local Depth}
\usage{
idld(z, data, beta, m, type, verbose)
}
\arguments{
\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{type}{the data type to apply the idld, "multivariate", "functional" or "multi_functional".}

\item{verbose}{if TRUE prints the algorithm progress.}

\item{Z}{data to apply depth. It should be an array of dimension (n,p,l) where l is the number of functional coordinates. 
Z[,,i] is a numeric matrix where each row represents a functional observation for i=1,...,l.}

\item{data_mf}{data on which depth is based. Same format than Z.}
}
\value{
A numeric vector object that contains the depth for each point.
}
\description{
Calculates the integrated dual local depth
}
