% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idld_multivariated.R
\name{idld_m}
\alias{idld_m}
\title{idld_m}
\arguments{
\item{Z}{data to apply depth. It should be a numeric matrix where each row represents an observation.}

\item{data_m}{data on which depth is based. It should be a numeric matrix where each row represents an observation.}

\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{verbose}{if TRUE prints the algorithm progress.}
}
\value{
A numeric vector that contains the depth for each point.
}
\description{
Calculates the integrated dual local depth for multivariate data
}
\examples{
library(mvnfast)
X = rmvn(90,  c(-2,-2), sigma=diag(rep(1,2)))
Y = rmvn(110, c(2,2), sigma=diag(rep(1,2)))
Z = rmvn(150, c(4,-4), sigma = rbind(c(2,0.8),c(0.8,1)))
W = rbind(X,Y,Z)
local_depth = idld(W,W,0.3,500,TRUE)
plot(W, pch=20)
local_depth_center_region = which(local_depth>quantile(local_depth,0.9))
points(W[local_depth_center_region,], col="blue", pch=20)

}
