% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idid_multifuncional.R
\name{idld_mf}
\alias{idld_mf}
\title{Functional Integrated Dual Local Depth}
\arguments{
\item{Z}{data to apply depth. It should be an array of dimension (n,p,l) where l is the number of functional coordinates. 
Z[,,i] is a numeric matrix where each row represents a functional observation for i=1,...,l.}

\item{data_f}{data on which depth is based. Same format than Z.}

\item{beta}{locality parameter between 0 and 1}

\item{m}{number of random projections}

\item{verbose}{if TRUE prints the algorithm progress.}
}
\value{
A numeric vector object that contains the depth for each point.
}
\description{
Calculates the integrated dual local depth for functional data and multivariate functional data.
}
\examples{
data(growth_domestic_product)
data(inflation_rate)
library(abind)
m_functional_data = abind(as.matrix(growth_domestic_product[,-1]/400), 
                          as.matrix(inflation_rate[,-1]), 
                          along = 3)
local_depth = idld_mf(m_functional_data, m_functional_data, 0.3, 500, TRUE)
}
