% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inflation_rate}
\alias{inflation_rate}
\title{Inflation Rate}
\format{
A data frame with 101 rows and 28 columns:
\describe{
  \item{country}{the country in which the inflation rate was measured.}
  \item{1990}{inflation rate in 1990 for each country.}
  ...
  \item{2016}{inflation rate in 2016 for each country.}
}
}
\source{
\url{https://www.imf.org/external/datamapper/NGDPD@WEO/OEMDC/ADVEC/WEOWORLD}
}
\usage{
inflation_rate
}
\description{
A dataset containing the inflation rate calculated for several countries from 1990 to 2016 based in 2016.
}
\keyword{datasets}
